+++ src/NEventLite/Core/Domain/Event.cs
@@ -8,11 +8,11 @@ namespace NEventLite.Core.Domain
         {
         }
 
-        protected Event(Guid id, Guid aggregateId, int targetVersion) : base(id, aggregateId, targetVersion)
+        protected Event(Guid id, Guid aggregateId, long targetVersion) : base(id, aggregateId, targetVersion)
         {
         }
 
-        protected Event(Guid id, Guid aggregateId, int targetVersion, int eventSchemaVersion, string correlationId) : 
+        protected Event(Guid id, Guid aggregateId, long targetVersion, int eventSchemaVersion, string correlationId) : 
             base(id, aggregateId, targetVersion, eventSchemaVersion, correlationId)
         {
         }
@@ -20,16 +20,16 @@ namespace NEventLite.Core.Domain
 
     public abstract class Event<TAggregate, TAggregateKey, TEventKey> : IEvent<TAggregate, TAggregateKey, TEventKey> where TAggregate : AggregateRoot<TAggregateKey, TEventKey>
     {
-        protected Event(TEventKey id, TAggregateKey aggregateId) : this(id, aggregateId, (int) StreamState.NoStream)
+        protected Event(TEventKey id, TAggregateKey aggregateId) : this(id, aggregateId, (long) StreamState.NoStream)
         {
         }
 
-        protected Event(TEventKey id, TAggregateKey aggregateId, int targetVersion) : this(id, aggregateId, targetVersion, 0,
+        protected Event(TEventKey id, TAggregateKey aggregateId, long targetVersion) : this(id, aggregateId, targetVersion, 0,
             string.Empty)
         {
         }
 
-        protected Event(TEventKey id, TAggregateKey aggregateId, int targetVersion, int eventSchemaVersion, string correlationId)
+        protected Event(TEventKey id, TAggregateKey aggregateId, long targetVersion, int eventSchemaVersion, string correlationId)
         {
             this.Id = id;
             this.AggregateId = aggregateId;
@@ -40,7 +40,7 @@ namespace NEventLite.Core.Domain
 
         public TEventKey Id { get; set; }
         public TAggregateKey AggregateId { get; set;  }
-        public int TargetVersion { get; set;  }
+        public long TargetVersion { get; set;  }
         public DateTimeOffset EventCommittedTimestamp { get; set; }
         public int EventSchemaVersion { get; set; }
         public string CorrelationId { get; set; }
