+++ src/Samples/NEventLite.Samples.Common/Domain/Schedule/Events/TodoCompletedEvent.cs
@@ -7,7 +7,7 @@ namespace NEventLite.Samples.Common.Domain.Schedule.Events
     {
         public Guid TodoId { get; set; }
 
-        public TodoCompletedEvent(Guid aggregateId, int targetVersion, Guid todoId) : base(Guid.NewGuid(), aggregateId, targetVersion)
+        public TodoCompletedEvent(Guid aggregateId, long targetVersion, Guid todoId) : base(Guid.NewGuid(), aggregateId, targetVersion)
         {
             TodoId = todoId;
         }
