+++ src/Samples/NEventLite.Samples.Common/Domain/Schedule/Events/TodoUpdatedEvent.cs
@@ -8,7 +8,7 @@ namespace NEventLite.Samples.Common.Domain.Schedule.Events
         public Guid TodoId { get; set; }
         public string Text { get; set; }
 
-        public TodoUpdatedEvent(Guid aggregateId, int targetVersion, Guid todoId, string text) : base(Guid.NewGuid(), aggregateId, targetVersion)
+        public TodoUpdatedEvent(Guid aggregateId, long targetVersion, Guid todoId, string text) : base(Guid.NewGuid(), aggregateId, targetVersion)
         {
             TodoId = todoId;
             Text = text;
